pyproject.toml
[project]
name = "proof-of-capture"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "c2pa-python>=0.32.3",
    "numpy>=2.4.4",
    "pillow>=12.2.0",
]
